sub-process: print the cmd when a capability is unsupported
Junio C Hamano <gitster@pobox.com> writes:

> I still think we would want to turn warning() to die(), but it
> probably is better to do so in a separate follow-up patch.  That
> will give us a good place to record the reason why the current "just
> call a warning() and pretend as if nothing bad happend" is wrong.

And here is such an update.  It seems that pretty much all comments
in the original thread were "warning is wrong--we should die here",
but nobody seems to have bothered following it through.

cf. <20170815111725.5d009b66@twelve2.svl.corp.google.com>

-- >8 --
Subject: [PATCH] subprocess: loudly die when subprocess asks for an unsupported capability

The handshake_capabilities() function first advertises the set of
capabilities it supports, so that the other side can pick and choose
which ones to use and ask us to enable in its response.  Then we
read the response that tells us what choice the other side made.  If
we saw something that we never advertised, that indicates one of two
things.  The other side, i.e. the "upgraded" filter, is not paying
attention of the capabilities advertisement, and asking something
its correct operation relies on, but we are not capable of giving
that unknown feature and operate without it, so after that point the
exchange of data is a garbage-in-garbage-out.  Or the other side
wanted to ask for one of the capabilities we advertised, but the
code has typo and their wish to enable a capability that its correct
operation relies on is not understood on this end.  The result is
the same garbage-in-garbage-out.

Instead of sweeping such a potential bug under the rug, die loudly
when we see a request for an unsupported capability in order to
force sloppily-written filter scripts to get corrected.

Signed-off-by: Junio C Hamano <gitster@pobox.com>