coccinelle: add rules to use strbuf_add() for adding buffers to strbufs
strbuf_addf() can be used with "%.*s" to add a buffer of a specific size
to a strbuf, but the resulting code is longer, may require casting the
length to int and is less efficient than simply calling strbuf_add().
Add Coccinelle rules for transforming calls of the former type to the
latter.

Signed-off-by: Rene Scharfe <l.s.r@web.de>