Documentation/git-config.txt: reword missleading sentence
Change the word "bla" to "section.variable", "bla" is a placeholder
for a variable name and it wasn't clear for everyone.
This change clarify it.

Change the appearance of 'git config section.variable {tilde}/' to
`git config section.variable {tilde}/` to harmonize it with
the rest of the file, this is a command line then the "`" are
necessary.

Replace "git-config" by "git config" because the command
is not "git-config".

See discussion at:
https://public-inbox.org/git/20171002061303.Horde.SL92grZCqTRV9oQkBFPELQ7@crashcourse.ca/

Signed-off-by: MOY Matthieu <matthieu.moy@univ-lyon1.fr>
Signed-off-by: Daniel Bensoussan <daniel.bensoussan--bohm@etu.univ-lyon1.fr>
Signed-off-by: Timothee Albertin <timothee.albertin@etu.univ-lyon1.fr>
Signed-off-by: Nathan Payre <second.payre@gmail.com>
Noticed-by: rpjday@crashcourse.ca